agent_system/requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.8
langchain-community>=0.0.20

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP and API
httpx>=0.25.0
requests>=2.31.0
tavily-python>=0.3.0

# Utilities
python-json-logger>=2.0.7
tiktoken>=0.5.0  # 保留用于向后兼容
transformers>=4.35.0  # Qwen tokenizer
torch>=2.0.0  # transformers 依赖

# Database and Cache
redis>=5.0.0  # Redis客户端
psycopg2-binary>=2.9.0  # PostgreSQL客户端
sqlalchemy>=2.0.0  # SQL工具库（可选）

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: Vector stores (for local testing only)
# chromadb>=0.4.0
# faiss-cpu>=1.7.4

# API
fastapi>=0.104.0
uvicorn>=0.24.0
uvicorn[standard]>=0.24.0

